+++ src/nationalarchives/components/error-summary/error-summary.stories.js
@@ -1,4 +1,5 @@
 import ErrorSummary from "./template.njk";
+import TextInput from "../text-input/template.njk";
 import macroOptions from "./macro-options.json";
 
 const argTypes = {
@@ -56,4 +57,39 @@ Standard.args = {
   ],
   disableAutoFocus: true,
   classes: "tna-error-summary--demo",
+};
+
+const FormTemplate = ({ items }) =>
+  `${ErrorSummary({
+    params: {
+      title: "There is a problem",
+      headingLevel: 2,
+      items: Array(items)
+        .fill("")
+        .map((item, index) => ({
+          text: `Error on field ${index + 1}`,
+          href: `#input-${index + 1}`,
+        })),
+      disableAutoFocus: true,
+      classes: "tna-error-summary--demo",
+    },
+  })}${Array(items)
+    .fill("")
+    .map((item, index) =>
+      TextInput({
+        params: {
+          label: `Field ${index + 1}`,
+          id: `input-${index + 1}`,
+          name: `input-${index + 1}`,
+          error: {
+            text: "There is a problem",
+          },
+        },
+      }),
+    )
+    .join("")}`;
+
+export const ScrollExample = FormTemplate.bind({});
+ScrollExample.args = {
+  items: 24,
 };